+++ src/main/java/com/almostreliable/kubeaa/extension/TileEntityEmpowererExtension.java
@@ -1,0 +1,8 @@
+package com.almostreliable.kubeaa.extension;
+
+import java.util.UUID;
+
+public interface TileEntityEmpowererExtension {
+
+    void kjsaa$setPlayer(UUID uuid);
+}